refactor: generalize owner validation
Created the validateOwnership function in util/serviceHelpers.js, which
is a generalized form of validateDocumentOwner, because the code could
be applied to education, experience, text snippets, and etc. to
validate their owners.

This refactor will be done in pieces, due to time.
validateDocumentOwner will be kept for now, because refactoring old
code for it will take a lot of time.