fix: hash password when updating it
When changing the password, the password was not being hashed.  This
caused users to not be able to sign in afterwards and passwords to be
plain text in the database.

This has now been corrected and tests for User.update, as well as
User.register, have been updated to check that the password in the
database is different than the supplied password.