chore: set authToken directly after getting it
The authToken state in App has been essentially removed, because there
is no foreseeable use for it.  It will only be used when making API
calls, therefore it can be set within the ResumeManagerApi Class and
immediately after registering or signing in a user.

In its stead, a user state has been created to store any user info that
might be used in the app.  Currently, the authToken is being stored as
its property, just in case it is needed later.  Though, it should be
removed in the future.

Written tests have been skipped due to time, but manual tests have been
done.