feat: implement document selection React component
Created a React component that displays provided documents in a modal
and allows a user to select which document to open or to create a new
one.  Note that a document ID of 0 is used to indicate creating a new
document.  Ensure that there is no document with ID = 0 in the database.

Written tests are skipped for now due to time, but basic manual tests
have been performed.