feat: implement initial document webpage
Created the initial parts of the webpage / React component for
displaying a document and its document selection.  This currently shows
document selection and allows for document creation.  Note that the
route protection against non-logged-in users is in the useEffect, but
this should later be put into a route protector.

The Document component has been added to the routes list.  The username
property is currently being passed down thru child props instead of a
context, to reduce time implementing a context.

createDocument and getDocuments functions are created in
ResumeManagerApi for convenience.

The document webpage link has been added to the NavBar.

The register, sign-in, and update account webpages now redirect to the
document webpage after submission.

Written tests have been skipped due to time, but basic manual tests
have been performed.