feat: implement add education React component
Created the React component that renders the form to add a new
education to the document and database.  This will appear after all
existing sections.

In the handleSubmit function, there is logic that'll remove fields with
empty Strings.  This is done here in the AddEducationCard React
component, because empty Strings are supposed to be null fields and
setting the correct field values from user inputs are supposed to be
done as early as possible.

At the end of AddEducationCard, the code is set up to reduce duplicate
code.  The fields variable must have required fields before optional
fields.  getRequiredIndicator, getPlaceholder, and getPattern helper
functions are used to set the correct Input attribute values
conditionally, such as for when fields are optional or fields that are
dates.  Note that the regex pattern in getPattern is a simple format
check and can not actually determine if a date actually exists.

The SectionCard React component has been updated to include rendering
form components that add section items.  It currently only has things
related to education.

addEducation function has been created in SectionCard, because it seems
to be the best place to put it, as it updates the section items that
are being displayed in SectionCard.  Although, it does not take any
inputs specific from SectionCard, so consider moving it elsewhere
later.  Contexts have been used as a result of having addEducation.

The sectionIdToSectionComponent is renamed to sectionComponents for
easier reading.  It also has been moved into SectionCard and closer to
where it is being used.  There was no need to have it outside of
SectionCard, as nothing else will be using it.

Implemented a method, addEducation, in ResumeManagerApi for
conveniently making an HTTP request.

Written tests have been skipped due to time, but basic manual tests
have been performed.