feat: implement delete section in front-end
In the SectionCard React component, added a delete button on section
cards and implemented its functionality in the deleteSection function.
This will allow removing sections in a document.

Also added in displaying of error messages, which currently is only
being used by deleteSection.  At the moment, error messages are being
rendered in the card header, because there isn't a better place.  At
the moment, the error messages disappear after a few seconds, because
otherwise, it will stay there forever.  Allowing the messages to be
closed by clicking is avoided to be consistent with the other
displaying of errors.

Also moved the card title into the card body, because it shouldn't be
outside of the body.

In ResumeManagerApi, created the deleteSection method as a convenience
from having to write out the URL.

Written tests have been skipped due to time, but basic manual tests
have been performed.