feat: save auth token in local storage and in API
Added functionality to store the authentication token on a user's
computer, which will forego the need to log in every time upon visiting
the website.

Updated ResumeManagerApi to make authToken private and added a getter
and setter for it.  This will allow adding extra code to store the
authToken in local storage, without having to create explicit methods.
In other words, authToken will be accessible from ResumeManagerApi like
it was a public property.

Since the authentication token will only be used by the API, it should
only be kept in it.  registerUser and signinUser in ResumeManagerApi
now return the decoded user data from the token, because returning the
token is not useful as it is not directly used elsewhere.

In the App React component, registerUser and signinUser have been
updated to directly store the returned user data, as opposed to
decoding the token and storing the user data, as a result of the API
not returning the token anymore.  Also, it was not necessary to
destructure existing user data, because there wouldn't be any when
registering and signing in.

In the App React component, the user state now initializes with the
user data, decoded from the authentication token, if the token exists
in local storage and has been stored in ResumeManagerApi.  This will
allow the React side of the app to see that a user has previously
logged in from a previous session.

Written tests have been skipped due to time, but basic manual tests
have been performed.