feat: update route protection in front-end
Logged-in users were able to access the registration and sign-in
webpages, which does not make sense as they are already logged in.

To keep route protection in one React component, RouteProtector has
been updated to redirect logged-in users to the "/document" route.
There is no need to modify the home page to welcome returning users and
have users redirect to that, as logged-in users would prefer to
immediately be on the document webpage.  The logic for non-logged-in
users has also been updated accordingly as well, because RouteProtector
is now being applied to all routes.

In RoutesList, the RouteProtector route now encompasses all routes.

Written tests have been skipped due to time, but basic manual tests
have been performed.