feat: implement delete education in front-end
Added functionality to allow deleting an education from the database if
deleting from the master resume, or removing an education from a
document (removing a document-education relationship) if the document
is not the master resume.  The document object in the app state is
locally updated instead of retrieving the document again, in order to
save on a network call.

Added the trash icon to the EducationCard React component to allow
users to do the deleting.  Also displays error messages next to the
trash icon, because currently there is no place to put it.

Added deleteEducation and removeEducationFromDocument methods in
ResumeManagerApi to help with this.

Written tests have been skipped due to time, but basic manual tests
have been performed.