refactor: remove username argument in API methods
In the front-end, in ResumeManagerApi, methods repeatedly had username
as their parameters.  This was originally done, because there was a
possibility that users might be able to interact with other users or
view other users' info.  This is no longer the case, and users will
only be able to interact with their own info.  Having ResumeManagerApi
methods with a username parameter, where the value will always be the
same, was unnecessary.

The username is now saved in ResumeManagerApi, when getting and setting
an authentication token.  This will help reduce duplicate code and
using a context.  The API methods have been updated with username
removed, and the calls to them have been updated accordingly.

Written tests have been skipped due to time, but basic manual tests
have been performed.