refactor: move cluttered code in SectionCard
The SectionCard React component was filled with code that belonged to
specific sections like education, or code that should have been in its
own component like for rendering section items (i.e. education).

The code for rendering the list of items, as well as the add/attach
item form, have been moved into SectionItemsList.  Code for passing
props has been removed, because they now exist in the components, which
will be explained below.

The functions for adding a new section item (i.e. education) and
attaching an existing section item to a document have been generalized
and moved to util/specificSectionsFuncs.js.  Switch statements are
used, because they seem to be the easier way of calling the specific
ResumeManagerApi methods.  The alternative could be to store the
methods in an Array, and then retrieve the appropriate method when
calling, but the end result might be about the same.  There is repeated
code at the end of the functions, which will later be refactored after
refactoring the similar functions for the section model.

AddEducationCard has been updated to import the new function,
addNewSectionItem, instead of receiving it as a React prop.  This
should help clear up some code.

AttachEducationCard has similarly been updated to use the new
attachSectionItem function.  The effect to get a user's educations in
the SectionCard component has been moved here, because this is where it
is only used now.

In SectionsList, sectionIdToDatabaseName has been moved into a common
file for storing data values.  SectionCard no longer needs to be passed
the list of items to render, as that's now being done in
SectionItemsList.

A file, commonData.js, has been created to hold common static data
values.

Functions related to section model will be refactored later.

Written tests have been skipped due to time, but basic manual tests
have been performed.