feat: implement delete document in front-end
In the Document React component, created a function called
deleteDocument, which will delete the currently viewing document from
the database and from the website view.  The documents state is locally
updated, instead of making another GET request, to save on a network
call.

Added in displaying of error messages in the Document React component.

In ResumeManagerApi, added a method called deleteDocument to help make
the HTTP API request to delete a document from the database.  This
method is created for convenience and helps reduce lengthy code.  Also
moved the getContactInfo method to keep consistent organization.

Written tests have been skipped due to time, but basic manual tests
have been performed.