feat: implement contact info in front-end
Created the ContactInfoCard React component, which will display the
contact information for a user and allowing editing it.

The Document React component has been updated to display
ContactInfoCard.  In the createDocument function, fixed a bug where
getting contact info for a new user, or someone does not have a contact
info entry in the database, would cause an alert to be displayed to the
user.  Getting contact info for a newly created document is not
extremely important, so the error thrown from
ResumeManagerApi.getContactInfo is now ignored.  In absence of contact
info, the ContactInfoCard will display empty fields and users can
re-input their contact info again.

Added updateContactInfo method in ResumeManagerApi.

Added a SVG of a pencil to use as the edit icon.

Also removed unnecessary curly braces for all renderings of the trash
icon.

Written tests have been skipped due to time, but basic manual tests
have been performed.