feat: create JSON schema for text snippet version
This schema was created, because when users want to attach a text
snippet to an education in a non-master resume, users need to give the
ID and version, which are the identifications for a text snippet.  ID
is checked in the URL params schema, but version is too complex to be
in the schema.  As a result, the version needs to be placed in the HTTP
request body, and a separate JSON schema is needed.

Also note that the version is a timestamp with timezone, so this schema
is essentially checking for a timestamp.