feat: implement AttachTextSnippetCard initials
Created front-end and back-end functionality for displaying the
Reactstrap card for attaching a text snippet to an experience, while
not on the master resume.  Only the display of the component and list
of text snippets for an experience have been implemented.

* Back-end:
In routes/experiences.js, implemented a route for retrieving all of the
text snippets for an experience.  All snippets are retrieved instead of
ones for a particular document, because when creating a new resume or
template, users would want to search through all achievements for an
experience.  There is no new route dedicated to text snippets, because
text snippets can only exist within experiences, skills, etc., like how
experiences can only exist within documents.

In services/experienceService.js, a service-level function, called
getTextSnippets, has been implemented to help validate ownership of an
experience.

In models/textSnippet.js, a new method, called getAllForExperience, has
been created to get text snippets for an experience.  The code has been
placed in this model, because it retrieves text snippets.  The existing
getAll method is not significantly changed, because that will break the
common tests.  It's best to leave getAll alone for now.

In models/textSnippet.js, a new private method, #allDbColsAsJs, has
been created to help with SQL JOIN statements in getAllForExperience,
because there are tables that have an attribute with the same name.
Therefore, each attribute needs to be prefixed with table names.  It
starts with a hash, because that is the newer convention of naming
private class variables.  The old _allDbColsAsJs has to be kept,
because existing code still uses it.  A refactoring in the future needs
to be done to remove it.

In models/textSnippet.js, the hardcoded class names, TextSnippet, used
for logs, have been replaced with dynamic retrievals.

* Front-end:
Created the AttachTextSnippetCard React component to allow the option
of attaching a text snippet to an experience while working on a
non-master resume.  It currently only retrieves and shows the available
text snippets to attach.  Note that the ResumeManagerApi method in the
useEffect block has been replaced with a passed down function.  This is
required to have AttachTextSnippetCard applicable to different
sections, instead of only experiences.

The TextSnippetsList React component has been updated to include
AttachTextSnippetCard.

In ExperienceCard, getAvailableTextSnippets has been implemented to get
all text snippets for an experience and is passed down to
TextSnippetsList.  It's place in ExperienceCard, because the
ResumeManagerApi method used in it is specific to experience.

In ExperienceCard, addTextSnippet and getAvailableTextSnippets have
switched to using React callbacks to improve performance.  This was
originally done when experimenting with using a context, but has been
kept afterwards.

In ResumeManagerApi, created the getTextSnippetsForExperience method to
call the URL to retrieve text snippets belonging to an experience.

Written tests have been skipped due to time, but basic manual tests
have been performed.