chore: update validateOwnership for text snippets
In util/serviceHelpers.js, validateOwnership was created with sections
(education, experience, etc.) in mind.  Since sections only had one ID
attribute, ID was set as one of the parameters.  Now, with the
implementation of text snippet routes, validateOwnership does not
immediately work, because text snippets also have a version attribute.

To solve this, the id parameter has been changed to being an Object
containing the identification properties.  For instance, for validating
an education, {id: 1} would be passed; for validating a text snippet,
{id: 1, version: '...'} would be passed.

All calls of validateOwnership have been updated to pass an Object
instead of the ID number.