feat: implement attaching text snippets
Created front-end and back-end functionality for attaching an existing
text snippet to an education.  This is for documents that are not the
master resume.

* Back-end:
In routes/experiences.js, implemented the route for creating a
experience-text snippet relationship.

In services/experienceService.js, implemented
createExperience_x_textSnippet to help verify ownership and to create
an Experience_X_Text_Snippet relationship.

The URL params schema is updated with also checking a text snippet ID
in the URL path parameters.

*Front-end:
In the AttachTextSnippetCard React component, implemented the
handleSubmit function.  This will get the necessary text snippet ID,
version, and Object and pass them along to the actual function that
will make the API call and update the local document state.  The
passing is necessary to keep AttachTextSnippetCard applicable to
sections other than experience.

In ExperienceCard, implemented the attachTextSnippet function to do the
actual work.  A shallow copy of document is adequate for updating the
document state, as nothing else is done to the document state and React
will re-render components soon after.

attachTextSnippet is passed down two children for now, because it is
simpler to implement and there is no context for passing data from
ExperienceCard to AttachTextSnippetCard yet.  Implementing the context
will be saved for a refactor.

In ResumeManagerApi, the relevant method,
attachTextSnippetToExperience, is created as a convenience from having
to write out the whole URL.

Written tests have been skipped due to time, but basic manual tests
have been performed.