feat: implement delete text snippet
Created front-end and back-end functionality for deleting a text
snippet.  Currently applied to experiences, because the rest of the
sections will not be implemented yet, but should be able to easily
apply it to them as well.

* Back-end:
Created a new file and route, routes/textSnippets.js, for deleting text
snippets.  This has its own file and route, because only the text
snippet ID and version are required to delete a text snippet.  Due to
the use of references and ON DELETE CASCADE constraints in the database
table schema, deleting a text snippet will remove any relationships it
has with any section.

Created a new service file and function, services/textSnippetService.js
and deleteTextSnippet, to help verify ownership of the text snippet to
be deleted.

* Front-end:
In the TextSnippetCard React component, implemented the
deleteTextSnippet function to allow deleting a text snippet if the
document is the master resume.  Part of the code is here, because it
can be generalized to all text snippets.  The rest of the code had to
be implemented in section-specific components and then passed down to
this component, because that code requires removing a text snippet from
an experience, skill, etc. in the document state.

Also in the TextSnippetCard React component, removed the data attribute
in the Card component in the returned JSX, because it was not needed.
The deleteTextSnippet function already has access to the ID and
version, since deleteTextSnippet is in the same component as the trash
icon.  Also made clicking on the trash icon run the deleteTextSnippet
function.

In the ExperienceCard React component, implemented
removeTextSnippetFromDocumentState, which will remove a text snippet
from the current experience that ExperienceCard is holding.  This code
had to be in ExperienceCard, else the experience data will have to be
passed down to and held onto by TextSnippetCard.

In the TextSnippetsList React component,
removeTextSnippetFromDocumentState is continued to be passed down to
the TextSnippetCard component.  This is minor prop drilling, done for
the sake of simplicity and time, and should be considered for
refactoring by using a context.

In ResumeManagerApi, implemented deleteTextSnippet as a convenience
method to reduce repeatedly writing out the URL.