fix: use new Object when updating state
In ExperienceCard, when updating the document state for addTextSnippet
and attachTextSnippet functions, the bullets property of experience did
not get a new Array if it already existed.  This meant that anything
that watches the bullets property would not see any changes and would
not cause necessary re-renders.

Instead of simply mutating the bullets Array, a new Array is now
created by using the spread syntax.  There is no current issue that
this solves, but this will help prevent any in the future.

Also, there was no need to create a documentCopy variable, as nothing
needed to be done to the document Object that was not done to the
experience Object.  The extraneous variable was likely a result of
adapted code from earlier that used structuredClone.

Written tests have been skipped due to time, but basic manual tests
have been performed.