chore: remove data attribute in HTML elements
In React components, data attributes were used to help find the
specific document, section, education, or experience to delete.  They
are not necessary after all, as the relevant components will have the
data and the delete functions contained within it.

As a result, uses of data attributes have been removed from the
Document, SectionCard, EducationCard, and ExperienceCard React
components to simplify the delete logic.

Also for SectionCard, EducationCard, and ExperienceCard, the deep clone
of document has been replaced with a shallow clone, to improve
performance.  This is possible, because not all of the document
properties are accessed and the rest of the React components are going
to be re-rendered soon after.

Written tests have been skipped due to time, but basic manual tests
have been performed.