fix: make removeTextSnippetFromDocumentState pure
The removeTextSnippetFromDocumentState function in ExperienceCard
caused a bug, during development for a future code that removes a text
snippet from an experience.  The bug removed the selected text snippet
from the view, along with the last text snippet in the list, because on
the second run of the function in strict mode, the text snippet was not
found, which then caused the supposedly found index to be -1.

This bug was caused by mutating the experience prop, which caused the
component to be impure.  To fix it, clones of all component inputs
(props and contexts) were first created before making changes to the
data.  Shallow copies were done for performance, which should still
work, but may possibly still cause issues in the future.  In short, the
main difference made was to make a copy of the experience Object,
instead of just the experience bullets.

The logic to find and remove the text snippet in the bullets Array has
also been replaced with an Array filter method.  This should shorten
the code, as well as help prevent the bug too.

The textSnippetVersion parameter has also been removed, because only
the text snippet ID is a part of the primary key in the database schema
for the experiences_x_text_snippets table.  This means that there can
only be one text snippet for a particular ID in a list of text
snippets, making the version is irrelevant.

Written tests have been skipped due to time, but basic manual tests
have been performed.