fix: allow only updating document name for master
In resume-manager-backend\services\documentService.test.js, the
updateDocument function was allowing the isTemplate and isLocked
properties to be updated for master resumes.  A master resume will
always be the source for other documents, so it always has to be a
template and unlocked.  For documents that are not master, any amount
of the properties can be updated.

The if condition has been corrected and overall code is also simplified.

ForbiddenError has been removed from the JSDoc description, because
only Errors thrown by the current code, and not depended code, should
be included.

Tests have been reworked accordingly.  For the first test in
documentService.test.js, 'Updates document properties for non-master
document.', an extra test input is given to test for no properties.