feat: implement edit education in front-end
Created the EducationForm React component, which contains a form to
input required and optional education fields.  It is an adapted,
generalized layout of the form in AddEducationCard and has been
developed in mind to be used for adding an education as well.  A lot of
the hardcoded data and common functions have been moved into
resume-manager-frontend\src\commonData.js.

In the EducationCard React component; EducationForm, the initial form
data for EducationForm, and a new edit button have been added to the
rendered output.  For the initialFormData variable, the list of fields
are populated from the EDUCATION_FIELDS constant, so that there is only
one place to get the actual fields that are going to used on the form.
Implemented the editEducation function, which will help call the
back-end API and update the React components.  Note that, currently,
only the master resume is allowed to edit, to keep the management of
info simple.

Implemented updateEducation in ResumeManagerApi to help make it easier
to call the back-end API.

Written tests have been skipped due to time, but basic manual tests
have been performed.