feat: implement back-end for updating experience
Created a new route for updating an experience.  This only updates
properties like title and dates and not the text
snippets/achievements/bullets.

The path was decided to be "/users/:username/experiences/:experienceId"
instead of
"/users/:username/documents/:documentId/experiences/:experienceId",
because it potentially conflicts with relationship routes, and the
document ID was only used to verify if it was a master document.  I
have decided now that it feels overreaching to have the back-end API
only allow master documents to make edits to experiences when
experiences are not tied to master documents.  This restriction seems
best performed in the front-end, as working on a specific document with
API calls is not really a thing.

A new schema, experienceUpdate, is created to help validate user
inputs.  It is copied from experienceNew, with the most significant
change being in the endDate pattern.  The pattern accepts an empty
String or a date, where an empty String means to clear out the endDate
value from the database.  This is only applied to and is necessary to
endDate, because endDate is the only optional field.

The associated service function, updateExperience, has been implemented
to help validate experience ownership.

Written tests have been skipped due to time, but basic manual tests
have been performed.