feat: implement edit experience in front-end
Implemented editing an experience (not including the text snippets /
bullets) by adding the GenericForm React component and an edit icon to
the rendered output of ExperienceCard.  The edit button only appears in
the master resume, because all additions, deletions, and updates should
be consolidated in the master resume.  Other resumes should only be
attaching and detaching information.  The edit form is integrated into
ExperienceCard, instead of being in something like a modal, because the
form should appear at where the experience is being displayed, so that
it's a seamless transition for users.  Also in ExperienceCard, the
editExperience function has been implemented to help make the API call
and update the document state.

In commonData.js, added the experience fields constants.

Implemented updateExperience in ResumeManagerApi to help make it easier
to call the back-end API.

Written tests have been skipped due to time, but basic manual tests
have been performed.