feat: implement edit text snippet in back-end
Created a new route in resume-manager-backend\routes\textSnippets.js
for editing/updating a text snippet.  The route path was chosen to be
shorter and direct, because a document, experience, skill, or project
is not relevant to the properties of a text snippet.  It is also
because this allows the route to be generalized to any section and
document.  Restrictions on whether to allow editing depending on
document or section type is best suited for the front-end.

Note that the text snippet version is passed in the body of an API
request, because it is a timestamp and may be too complicated to be in
the route path.

Created the textSnippetUpdate schema, which is copied over from
textSnippetNew.

Created the associated service function, updateTextSnippet, to help
with validating text snippet ownership.

Written tests have been skipped due to time, but basic manual tests
have been performed.