refactor: use model methods in exp - snippet tests
For the integration tests for Experience_X_Text_Snippet, the outermost
beforeAll block was directly inserting test data into the database
through SQL statements.  The code was lengthy and did not increment
database entry IDs.  The code has now been replaced, but this has also
caused the requirement of storing the returned data from the model
methods.  This means that test data can only be accessed in setup
blocks or within tests and can not be used in the parameters for
parameterized tests.

In the describe blocks for each method, but before tests, the
"dataToAdd" and "existingData" variables had to be assigned values in
beforeAll blocks, because the values only existed after running the
beforeAll block of the outermost describe block.

In the getAll section, the test 'Get all of %i
experience_x_text_snippet(s) for a document.' could no longer be
parameterized, because the parameters were using values created during
beforeAll or beforeEach blocks and Jest evaluates the parameters before
any setup blocks.

Also changed some variable names to follow JavaScript naming convention.