refactor: use React context for text snippet funcs
In the front-end, in the TextSnippetsList React component, the list of
functions received as props have been moved into a React context to
avoid prop drilling.

A context called TextSnippetContext has been created in
resume-manager-frontend\src\contexts.jsx to allow for this refactor.

In ExperienceCard, functions have been removed as props for
TextSnippetsList and placed in a React context provider.  There is no
worry about mixing and getting the wrong text snippet contexts between
sections and section items, because the provider only wraps around
TextSnippetsList, and because an instance of TextSnippetsList and its
children only belongs to an experience instance.  The functions are
placed in a memo before being given to the context provider for
performance reasons.

Also in ExperienceCard, functions have been moved around to follow the
CRUD order.

In TextSnippetCard, AddTextSnippetCard, and AttachTextSnippetCard,
functions that were passed down as props are now being retrieved from a
context.

Written tests have been skipped due to time, but basic manual tests
have been performed.