feat: implement repositioning sections front-end
Implemented the functionality to allow dragging and dropping sections
and saving the new positions into the back-end.  Implemented code have
been adapted from examples from @hello-pangea/dnd
(https://github.com/hello-pangea/dnd/blob/HEAD/docs/about/examples.md).

Installed @hello-pangea/dnd package to add drag and drop functionality.

In SectionsList, added DragDropContext from @hello-pangea/dnd to
rendered output.  @hello-pangea/dnd suggests wrapping the whole app in
DragDropContext, but it is only applied to rendering SectionCards to
limit its scope.  Also added Droppable to make the list of SectionCards
a drop zone.  The index of the SectionCards need to be passed into
SectionCard, because @hello-pangea/dnd Draggables requires it.

In SectionsList, implemented the onDragEnd function to handle the end
of a drag.  This only deals with the dragging and what items were
dragged, in order to separate out responsibilities.

In SectionsList, implemented the repositionSections function, which
will update the document state with the new positions of the sections
and send an API request to save the new positions.  If the API request
fails, the original order of the sections, before the current set of
reorderings, will be restored.  The API request is also delayed, so
that if a user makes multiple quick reorderings, there won't be
multiple network calls.

In SectionCard, added Draggable from @hello-pangea/dnd to designate the
SectionCard as something that can be dragged.  The drag handle is a set
of dots, which is used to limit where on a section can initiate a drag.

In ResumeManagerApi, implemented the repositionSections method to make
it easier to call the back-end API.

Written tests have been skipped due to time, but basic manual tests
have been performed.