chore: move route for updating text snippet
The route to update a text snippet was in the file
resume-manager-backend\routes\textSnippets.js.  The associated service
function had code that was specific to experience.  This made the route
and service appear to be applicable to all text snippets.  As a result,
the route and service have been moved to
resume-manager-backend\routes\experiences_x_textSnippets.js and
resume-manager-backend\services\experienceXTextSnippetService.js,
because text snippets are specific to either experiences, skills, or
projects.

Written tests have been skipped due to time, but basic manual tests
have been performed.