feat: implement reposition text snippets front-end
Added front-end functionality for reordering text snippets for
experiences.

In TextSnippetsList, added the Droppable component to the rendered
output.  Since TextSnippetsList is used multiple times across different
experiences, the droppableId and type properties for Droppable had to
be unique, or else @hello-pangea/dnd will cause conflicts and allow
text snippets to interact across experiences.  To make the properties
unique, they had to be passed in as arguments, where the argument
values are uniquely created from the parent component.  Also passed the
indices for TextSnippetCards into them, for use in @hello-pangea/dnd
Draggables.

In TextSnippetCard, added the Draggable component and the dots drag
handle to the rendered output.

In ExperienceCard, added the necessary droppableId and type properties
for use in TextSnippetsList's Droppable, which had to be unique with
respect to every experience.

In SectionsList, created the repositionExperienceTextSnippet function
to reposition text snippets in an experience, both visually and in the
back-end, and added handling to choose that in the onDragEnd function.
The code for repositionExperienceTextSnippet could not be adapted into
repositionHelper, because it dealt with an Array within an experience
in the document Object state, while repositionHelper dealt with an
Array within the document Object state.  In onDragEnd, since
repositioning text snippets are specific to an experience, the
experience ID had to be extracted out of the type property of
Droppables; there was no better convenient alternative.  Also moved and
renamed the common timeDelay variable into a higher scope and to
repositionTimeDelay.

For repositionExperienceTextSnippet, the choice of using Maps for the
React refs, repositionExperienceTextSnippetsTimeoutIdsRef and
oldExperienceTextSnippetsRef, was made because the only identification
between text snippet lists was the experience ID that they belonged
to.  Using an Array is not practical as the experience ID could be any
random number and there would be no way of finding the associated
values.

In ResumeManagerApi, implemented the repositionExperienceTextSnippets
method to make it easier to call the back-end API.

Written tests have been skipped due to time, but basic manual tests
have been performed.