feat: improve contact info JSON schema
Set the minimum length for full name to 2, so that it is something more
sensible.

Added a regex pattern for the phone number to have basic matching
against all phone numbers around the world.  This checks for at least 5
digits, which should reasonably be the minimum, at most one left and
right parentheses, allows for +, -, (, and ) symbols, and at most 20
characters, which should reasonably be the max amount of characters.

Added and updated appropriate test cases for the new phone number
pattern and the change to full name.