fix: require properties when updating contact info
The contact info JSON schema was allowing an empty Object to be
acceptable for creating a new or updating a contact info.  As this
would not make sense, the JSON schema now includes the "anyOf" keyword
to require at least one property to be present.

The related tests have been updated.