refactor: move update user route code into service
For the PATCH route that updates a user's account info, such as
password, the code logic was written in the route file, which does not
separate concerns.

The code has now been moved into a new userService file and updateUser
service function.  New unit tests have been created for the service
function.  Note that there is one test that is temporary, which tests
the else condition in the service code.  This test would not need to
exist if there were more account settings/properties to change.

Also, in relation to the JSON schema, the part of the code that checks
the existence of an old password if a new one is given has been
removed, because that functionality is now performed by the
userUpdateSchema JSON schema.  The users PATCH route tests have been
updated accordingly.