chore: reword error when updating user password
When changing a user's password, if the old password is incorrect, the
error message the user receives was 'Invalid username/password.'  This
could be confusing as users would not know if this applied to the old
or new password.

In resume-manager-backend\services\userService.js -> updateUser, the
SigninError thrown by User.signin is now caught and reworded to be
clearer.  Note that any other Error thrown by User.signin continues
unhindered.