refactor: break up DocumentSelect code
In the front-end, in the DocumentSelect React component, the rendered
output contained too much code to render the HTML.  To make the
returned output simpler, the form group for the document option and the
add-new-resume option have been moved into subcomponents,
DocumentOption and AddNewResumeOption.  This also improves performance,
because React can now skip rendering the subcomponents if nothing for
them changed.

React was also not handling the selected/checked option state, which
caused a bug, after the refactoring, where an option had to be clicked
twice for it to be selected.  This is fixed by setting the "checked"
property for the Input components.

The documentId state has been renamed to selectedOption, to better
indicate that it holds the selected option.

Written tests have been skipped due to time, but basic manual tests
have been performed.