chore: disable isTemplate input for new document
Since the template feature has not been implemented, the option, for
allowing users to set whether a new document should be a template,
should be disabled.  Temporary minor adjustments to back-end code have
been made, so that it would be easy to reverse these changes if the
feature gets implemented, and to reduce time and app-breaking bugs.

In the documentNewSchema, only the requirement for isTemplate is
removed, so that it is easier to get back the requirement.  This is
better than having to rewrite code related to isTemplate.

For the POST route for creating a new document, in the call to
Document.add, the argument now overrides any isTemplate value with
false.  This saves from having to make bigger changes.

The route to update document properties do not need to be changed, as
the isTemplate property is optional.  Users accessing this app through
the front-end will not notice there was an option for setting a
document as a template.  Clients accessing the back-end API will still
be able to set a document as a template, but it will do nothing.

Written tests have been skipped due to time, but basic manual tests
have been performed.