chore: remove setting document template and lock
Since the template and locking document features have not been
implemented, the forms displayed to users have been changed to only
show the document name.  This is done by having the initialFormData
value, that is passed into DocumentForm in the Document component,
contain only the documentName property.  The code to have the template
and lock options have been commented out, so that it can serve as a
reminder and also makes it easier to re-implement them.

There is no change necessary for the back-end regarding locking, as
that is allowed only when editing documents and as the isLocked
property to send to the back-end is optional.

The changes for the back-end regarding templates have already been made.

Written tests have been skipped due to time, but basic manual tests
have been performed.