perf: limit GET calls when attaching in front-end
In the SectionsList React component, an API call to retrieve the
available sections that can be attached to the document was made every
time it was mounted, regardless if a user was going to attach one.

The call (ResumeManagerApi.getSections) has been moved into
AddSectionCard and only occurs once, when the component is toggled.
The availableSections state now initializes with null, to differentiate
between not making a call yet and receiving no sections after making
one.  Note that the section Object is now passed from within
AddSectionCard.handleSubmit to attachSection and did not need to be
manually created again.  It is instead extracted from the available
sections retrieved from the API call.

In SectionsList, since the API has been moved, the useEffect block is
no longer needed and is removed. The documentClone has been replaced
with a shallow clone for better performance, and the interactions with
its sections property have been updated to make shallow copies to
indicate to any potential future React component that sections has been
changed.

Names containing "addSection" have also been changed to "attachSection"
to better indicate that a new kind of section is not being created, but
a section is being attached.

Written tests have been skipped due to time, but basic manual tests
have been performed.