feat(front-end): stylize document card headers
Added three elements to card headers, so that icons can be evenly and
properly distributed.  The griping dots icon can easily be centered be
placing it alone in the middle element.  The pencil and trash icons can
easily be positioned to the right be placing them in the last element
and making that element a flex container.  This strategy does require
having empty elements in the React component rendered outputs.  span
HTML elements were used, because they helped keep the items in one
single row.

Placed the edit and trash icon images in span HTML elements, so that a
circle background on hover can be added.

Used better, descriptive HTML tags for Card, CardHeader, and CardBody
Reactstrap components instead of the default div tag.

Replaced the three-dots icon with the standard gripping dots icon.