feat(front-end): stylize general document space
In the Document React component, the "Select Document" and edit buttons
utilized Reactstrap's prop for adding color instead of adding a class,
in order to keep consistency with Reactstrap and Bootstrap.  The
"primary" color option should be utilized for future color choices.

In Document.css, the CSS properties for the Document__page class is
meant to mimic a page of paper.

Overrode some Bootstrap properties regarding the primary color of
buttons by importing the "bootstrapOverride" CSS file after importing
the Bootstrap CSS within the main.jsx file.  Overriding the Bootstrap
CSS this way helps consolidate any future overrides in one location.
Note that in the overriding CSS, "#root" had to be used because the
custom CSS properties still did not have precedence or more specificity
than the Bootstrap CSS.

In index.css, removed unused CSS properties.

In the App React component, removed the extra div element, as that
provided no benefit.

In App.css, removed the unused margin property and set the width to
100% so that the whole webpage looks like a workspace.