fix: change API call for updating text snippet
Originally, the API endpoint for updating a text snippet was generic
and intended to be used for any resume section (experience, skill,
project).  This did not work, as the changes from the updated text
snippet needed to be propagated to the relationships between the text
snippet and the section it belonged to, because each update to a
snippet actually creates a new text snippet.  The back-end API was
already changed to account for this by being specific to each section
(currently only experience), but the front-end did not.

In ResumeManagerApi, updateTextSnippet has been modified and renamed to
be specific to experiences.

In TextSnippetCard, the editTextSnippet function no longer can call the
relevant ResumeManagerApi function, because it is now specific to
experiences, skills, projects, etc..  As a result, the network call and
updating the document state is now abstracted to the ExperienceCard
React component.

In ExperienceCard, the updateTextSnippet function has been created to
contain the logic for updating an experience's text snippet in the
back-end and in the document state.  The
replaceTextSnippetInDocumentState function no longer needed to be
passed to the text snippet context, because only updateTextSnippet uses
it, but the function is kept in case it might be used again.

Written tests have been skipped due to time, but basic manual tests
have been performed.