feat(front-end): use primary colors for Buttons
Set the color attribute for all Button Reactstrap components to
"primary" to utilize the overridden Bootstrap "btn-primary" class.

In bootstrapOverride.css, changed the CSS selectors to start with
"body" instead of "#root", because the color scheme needed to be
applied to the whole app.  Modals were not using the overridden CSS
properties because they are not children of the "#root" element.  Box
shadows have also been applied to all Buttons.