fix: reset open forms when changing documents
There was a bug where if add or attach forms were open in a document,
it will remain open when switching to another document.  This occurs
because React did not unmount those components.

This is fixed by setting the document state to null, or in other words,
clear out the document, before loading another document.

Written tests have been skipped due to time, but basic manual tests
have been performed.