feat(front-end): stylize document select component
Added CSS styling to the DocumentSelect React component.

In the DocumentSelect React component, set some Reactstrap components
to use certain semantic tags for better accuracy, instead of the
default "div" element.  Placed the list of documents in an unordered
list ("ul" element), because it is a list.  The converted creation and
updated timestamps utilize useMemos, because using JavaScript's Date's
toLocaleString method is expensive.  Removed displaying of whether a
document is a master, a template, or is locked, because they are not
implemented or no longer relevant.  The displaying of the master
document separate from the rest will need to be done in a future commit.

Created a CSS variable in the ":root" selector in index.css to hold the
primary color of the app.  A later refactor will be needed to remove
all of the hardcoded values across all CSS files.

Created a utility function in conversions.js called
generateDateTimeString to help convert a timestamp in ISO format into a
user-friendly format.  Also renamed the parameter of and reworded the
description of generateDateString.

Written tests have been skipped due to time, but basic manual tests
have been performed.